requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.1
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
exceptiongroup==1.3.0
frozenlist==1.7.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
magic-filter==1.0.12
multidict==6.5.0
packaging==25.0
pillow==11.2.1
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.1.1
urllib3==2.4.0
yarl==1.20.1
